Fix state serialization bug with pandas.DataFrame. (#126)
To fix this issue we add custom JSON encoding and decoding logic.

For most accurate serialization/deserialization, we use Pandas's built
in JSON serialization (via to_json) with the "table" strategy. This is
more verbose, but provides enough metadata to deserialize as close as
possible to the original DataFrame.

To deserialize we use `pandas.read_json` with `orient="table"`.

One difference I've noticed so far is that pandas.NA becomes np.nan.