Fix cutoff content for history prompt list / code diff viewer (#916)
I noticed that the Angular dialog sets a 65vh. But our code is
using 80vh. So I set it to 65vh to match.

One consideration though is whether this 65vh changes dynamically or
not. I did try adjusting viewport sizes in Chrome dev tools and it
seemed like it stayed at 65vh but not sure.

Closes #908